refactor: add comments and READ md
Add comments to service layer and modifie read md file
