+++ logiweb/src/main/java/com/logiweb/avaji/services/api/profile/OrderDetailsService.java
@@ -6,8 +6,38 @@ import com.logiweb.avaji.exceptions.ShiftValidationException;
 
 import java.util.List;
 
+/**
+ * Order Details service manipulate with OrderDetails entity in database.
+ * This service read OrderDetails by driverId, change cargo status of order which related to current OrderDetails
+ * and change city by and calculate new remaining path.
+ *
+ */
 public interface OrderDetailsService {
+
+    /**
+     * Read OrderDetailsDTO by driverId
+     *
+     * @param driverId
+     * @return OrderDetailsDTO
+     */
     OrderDetailsDTO readOrderDetails(long driverId);
+
+    /**
+     * Update Cargos by cargoIds of Order which related to current OrderDetails.
+     * Before updating DriverStatus to LOAD_UNLOAD_WORK and activate shift active status if
+     * it be inactive. If all Cargo status will be DELIVERED, change Order status to complete.
+     *
+     * @param driverId by finding OrderDetails
+     * @param cargoIds which will be changed
+     * @throws ShiftValidationException if DriverStatus and shift active status have conflict
+     */
     void updateOrderByCargoStatus(long driverId, List<Long> cargoIds) throws ShiftValidationException;
+
+    /**
+     * Change city to next by calculate remaining path
+     *
+     * @param orderId
+     * @param driverId
+     */
     void changeCity(long orderId, long driverId);
 }
