+++ logiweb/src/main/java/com/logiweb/avaji/services/implementetions/path/PathDetailsServiceImpl.java
@@ -1,6 +1,5 @@
 package com.logiweb.avaji.services.implementetions.path;
 
-import com.logiweb.avaji.daos.TruckDAO;
 import com.logiweb.avaji.dtos.*;
 
 import com.logiweb.avaji.services.api.map.CountryMapService;
@@ -22,14 +21,11 @@ public class PathDetailsServiceImpl implements PathDetailsService {
 
     private static final Logger logger = LogManager.getLogger(PathDetailsServiceImpl.class);
 
-    private final TruckDAO truckDAO;
     private final CountryMap countryMap;
 
     @Autowired
-    public PathDetailsServiceImpl(CountryMapService countryMapService,
-                                  TruckDAO truckDAO) {
+    public PathDetailsServiceImpl(CountryMapService countryMapService) {
         this.countryMap = new CountryMap(countryMapService);
-        this.truckDAO = truckDAO;
     }
 
     @Override
@@ -94,13 +90,7 @@ public class PathDetailsServiceImpl implements PathDetailsService {
         return from.until(to, ChronoUnit.HOURS);
     }
 
-    @Override
-    public int calculateFreeSpaceInShift(long orderId) {
-        TruckDTO truck = truckDAO.findTruckByOrderId(orderId);
-        int currentSize = (int) truckDAO.countDriversOfTruck(truck.getTruckId());
-        int size = truck.getShiftSize();
-        return (size - currentSize);
-    }
+
 
 
     @Override
